refactor(frontend): split demo data by domain
