+++ frontend/src/features/lms/data/demo-auth-courses.ts
@@ -1,0 +1,104 @@
+import type { AuthUser, CourseCard } from '@myway/shared';
+
+export const demoUsers: AuthUser[] = [
+  {
+    id: 'usr_demo_admin',
+    name: '김운영',
+    email: 'admin@mywayclass.dev',
+    role: 'ADMIN',
+    department: '플랫폼 운영팀',
+    bio: '강의, 사용자, AI 흐름을 모두 관리하는 운영자',
+  },
+  {
+    id: 'usr_demo_instructor_1',
+    name: '이서준',
+    email: 'instructor1@mywayclass.dev',
+    role: 'INSTRUCTOR',
+    department: '경영학과',
+    bio: '데이터와 비즈니스 전략을 함께 다루는 강사',
+  },
+  {
+    id: 'usr_demo_instructor_2',
+    name: '박지현',
+    email: 'instructor2@mywayclass.dev',
+    role: 'INSTRUCTOR',
+    department: '컴퓨터공학과',
+    bio: '실습형 프로젝트와 AI 워크플로를 설계하는 강사',
+  },
+  {
+    id: 'usr_demo_student_1',
+    name: '최민수',
+    email: 'student1@mywayclass.dev',
+    role: 'STUDENT',
+    department: '경영학부',
+    bio: '수강한 강의의 요약과 숏폼을 바로 보는 학습자',
+  },
+  {
+    id: 'usr_demo_student_2',
+    name: '윤하린',
+    email: 'student2@mywayclass.dev',
+    role: 'STUDENT',
+    department: '소프트웨어학과',
+    bio: '전사와 타임스탬프로 빠르게 복습하는 학습자',
+  },
+];
+
+export const demoCourses: CourseCard[] = [
+  {
+    id: 'crs_react_01',
+    instructor_id: 'usr_demo_instructor_2',
+    title: 'AI 서비스 설계 입문',
+    description: '전사, 요약, 숏폼, 챗봇을 한 흐름으로 묶는 제품 구조를 실습합니다.',
+    category: 'AI',
+    difficulty: 'intermediate',
+    is_published: true,
+    tags: ['AI', '제품', '오케스트레이션'],
+    instructor_name: '박지현',
+    lecture_count: 6,
+    enrolled: true,
+    progress_percent: 68,
+    completed_lectures: 4,
+    thumbnail_palette: 'indigo',
+    rating: 4.9,
+    student_count: 128,
+    total_duration_minutes: 480,
+  },
+  {
+    id: 'crs_java_01',
+    instructor_id: 'usr_demo_instructor_1',
+    title: '데이터 기반 의사결정',
+    description: '지표 설계, 대시보드 읽기, 실무 리포트 해석을 다룹니다.',
+    category: '경영',
+    difficulty: 'beginner',
+    is_published: true,
+    tags: ['경영', '지표', '리포트'],
+    instructor_name: '이서준',
+    lecture_count: 5,
+    enrolled: true,
+    progress_percent: 42,
+    completed_lectures: 2,
+    thumbnail_palette: 'emerald',
+    rating: 4.8,
+    student_count: 96,
+    total_duration_minutes: 360,
+  },
+  {
+    id: 'crs_demo_shortform',
+    instructor_id: 'usr_demo_instructor_2',
+    title: '숏폼 제작 워크플로',
+    description: '강의 하이라이트를 자동 추출하고, 바로 숏폼으로 만드는 과정입니다.',
+    category: '콘텐츠',
+    difficulty: 'intermediate',
+    is_published: true,
+    tags: ['숏폼', '편집', 'AI'],
+    instructor_name: '박지현',
+    lecture_count: 4,
+    enrolled: false,
+    progress_percent: 0,
+    completed_lectures: 0,
+    thumbnail_palette: 'violet',
+    rating: 4.7,
+    student_count: 74,
+    total_duration_minutes: 220,
+  },
+];
