+++ frontend/src/features/lms/data/demo.ts
@@ -1,395 +1,9 @@
-import type {
-  AIInsights,
-  AudioExtraction,
-  AuthUser,
-  CourseCard,
-  CourseDetail,
-  Dashboard,
-  LectureDetail,
-  LecturePipeline,
-  LectureTranscript,
-  MediaProcessorHealth,
-} from '@myway/shared';
-
-const now = new Date().toISOString();
-
-export const demoUsers: AuthUser[] = [
-  {
-    id: 'usr_demo_admin',
-    name: '김운영',
-    email: 'admin@mywayclass.dev',
-    role: 'ADMIN',
-    department: '플랫폼 운영팀',
-    bio: '강의, 사용자, AI 흐름을 모두 관리하는 운영자',
-  },
-  {
-    id: 'usr_demo_instructor_1',
-    name: '이서준',
-    email: 'instructor1@mywayclass.dev',
-    role: 'INSTRUCTOR',
-    department: '경영학과',
-    bio: '데이터와 비즈니스 전략을 함께 다루는 강사',
-  },
-  {
-    id: 'usr_demo_instructor_2',
-    name: '박지현',
-    email: 'instructor2@mywayclass.dev',
-    role: 'INSTRUCTOR',
-    department: '컴퓨터공학과',
-    bio: '실습형 프로젝트와 AI 워크플로를 설계하는 강사',
-  },
-  {
-    id: 'usr_demo_student_1',
-    name: '최민수',
-    email: 'student1@mywayclass.dev',
-    role: 'STUDENT',
-    department: '경영학부',
-    bio: '수강한 강의의 요약과 숏폼을 바로 보는 학습자',
-  },
-  {
-    id: 'usr_demo_student_2',
-    name: '윤하린',
-    email: 'student2@mywayclass.dev',
-    role: 'STUDENT',
-    department: '소프트웨어학과',
-    bio: '전사와 타임스탬프로 빠르게 복습하는 학습자',
-  },
-];
-
-export const demoCourses: CourseCard[] = [
-  {
-    id: 'crs_react_01',
-    instructor_id: 'usr_demo_instructor_2',
-    title: 'AI 서비스 설계 입문',
-    description: '전사, 요약, 숏폼, 챗봇을 한 흐름으로 묶는 제품 구조를 실습합니다.',
-    category: 'AI',
-    difficulty: 'intermediate',
-    is_published: true,
-    tags: ['AI', '제품', '오케스트레이션'],
-    instructor_name: '박지현',
-    lecture_count: 6,
-    enrolled: true,
-    progress_percent: 68,
-    completed_lectures: 4,
-    thumbnail_palette: 'indigo',
-    rating: 4.9,
-    student_count: 128,
-    total_duration_minutes: 480,
-  },
-  {
-    id: 'crs_java_01',
-    instructor_id: 'usr_demo_instructor_1',
-    title: '데이터 기반 의사결정',
-    description: '지표 설계, 대시보드 읽기, 실무 리포트 해석을 다룹니다.',
-    category: '경영',
-    difficulty: 'beginner',
-    is_published: true,
-    tags: ['경영', '지표', '리포트'],
-    instructor_name: '이서준',
-    lecture_count: 5,
-    enrolled: true,
-    progress_percent: 42,
-    completed_lectures: 2,
-    thumbnail_palette: 'emerald',
-    rating: 4.8,
-    student_count: 96,
-    total_duration_minutes: 360,
-  },
-  {
-    id: 'crs_demo_shortform',
-    instructor_id: 'usr_demo_instructor_2',
-    title: '숏폼 제작 워크플로',
-    description: '강의 하이라이트를 자동 추출하고, 바로 숏폼으로 만드는 과정입니다.',
-    category: '콘텐츠',
-    difficulty: 'intermediate',
-    is_published: true,
-    tags: ['숏폼', '편집', 'AI'],
-    instructor_name: '박지현',
-    lecture_count: 4,
-    enrolled: false,
-    progress_percent: 0,
-    completed_lectures: 0,
-    thumbnail_palette: 'violet',
-    rating: 4.7,
-    student_count: 74,
-    total_duration_minutes: 220,
-  },
-];
-
-export const demoCourseDetail: CourseDetail = {
-  ...demoCourses[0],
-  lectures: [
-    {
-      id: 'lec_react_01',
-      course_id: 'crs_react_01',
-      title: 'AI 오케스트레이션 개요',
-      order_index: 0,
-      week_number: 1,
-      session_number: 1,
-      content_type: 'video',
-      content_text: '제품 아키텍처와 AI 흐름을 연결하는 핵심 개념을 설명합니다.',
-      duration_minutes: 24,
-      is_published: true,
-      is_completed: true,
-      video_url: 'https://cdn.mywayclass.dev/demo/ai-orchestration-intro.mp4',
-      video_asset_key: 'media/demo/ai-orchestration-intro.mp4',
-    },
-    {
-      id: 'lec_react_02',
-      course_id: 'crs_react_01',
-      title: '전사와 타임스탬프',
-      order_index: 1,
-      week_number: 1,
-      session_number: 2,
-      content_type: 'video',
-      content_text: '타임라인과 노트 구조를 통해 복습용 맥락을 만드는 방법을 다룹니다.',
-      duration_minutes: 30,
-      is_published: true,
-      video_url: 'https://cdn.mywayclass.dev/demo/transcript-timeline.mp4',
-      video_asset_key: 'media/demo/transcript-timeline.mp4',
-    },
-    {
-      id: 'lec_java_01',
-      course_id: 'crs_react_01',
-      title: '숏폼 생성과 배포',
-      order_index: 2,
-      week_number: 2,
-      session_number: 1,
-      content_type: 'video',
-      content_text: '강의 하이라이트를 짧은 영상으로 재가공하는 배포 흐름을 실습합니다.',
-      duration_minutes: 28,
-      is_published: true,
-      video_url: 'https://cdn.mywayclass.dev/demo/shortform-export.mp4',
-      video_asset_key: 'media/demo/shortform-export.mp4',
-    },
-  ],
-  materials: [
-    {
-      id: 'mat_demo_ai_1',
-      course_id: 'crs_react_01',
-      title: 'AI 서비스 설계 체크리스트',
-      summary: '전사, 요약, 챗봇, 숏폼 연결 시 확인해야 할 항목',
-      file_name: 'ai-service-checklist.pdf',
-      uploaded_by: 'usr_demo_instructor_2',
-      uploaded_at: now,
-    },
-    {
-      id: 'mat_demo_ai_2',
-      course_id: 'crs_react_01',
-      title: '오케스트레이션 설계 노트',
-      summary: '프론트, 백엔드, shared 모듈의 책임 분리 예시',
-      file_name: 'orchestration-notes.md',
-      uploaded_by: 'usr_demo_instructor_2',
-      uploaded_at: now,
-    },
-  ],
-  notices: [
-    {
-      id: 'ntc_demo_ai_1',
-      course_id: 'crs_react_01',
-      title: '첫 주차 수업 안내',
-      content: '이번 주는 전사, 타임스탬프, 숏폼 생성의 전체 흐름을 실습합니다.',
-      pinned: true,
-      author_id: 'usr_demo_instructor_2',
-      created_at: now,
-    },
-  ],
-};
-
-export const demoLectureDetail: LectureDetail = {
-  ...demoCourseDetail.lectures[0],
-  course_title: demoCourseDetail.title,
-  course_instructor: demoCourseDetail.instructor_name,
-  previous_lecture_id: undefined,
-  next_lecture_id: demoCourseDetail.lectures[1].id,
-  is_completed: true,
-  video_url: demoCourseDetail.lectures[0].video_url ?? 'https://cdn.mywayclass.dev/demo/ai-orchestration-intro.mp4',
-  transcript_excerpt: 'AI는 단순 자동완성이 아니라, 제품 흐름을 따라 기능을 연결하는 실행 파트너로 다뤄야 합니다.',
-  keywords: ['오케스트레이션', '전사', '숏폼', '챗봇'],
-};
-
-export const demoDashboard: Dashboard = {
-  learner_name: '데모 수강생',
-  role: 'ADMIN',
-  total_courses: demoCourses.length,
-  active_enrollments: 3,
-  average_progress: 56,
-  courses: demoCourses,
-  stats: [
-    { id: 'enrollments', label: '활성 수강', value: '3', hint: '현재 운영 중인 수강 등록', icon: 'ri-user-follow-line', tone: 'emerald' },
-    { id: 'ai', label: 'AI 요청', value: '148', hint: '최근 AI 사용량', icon: 'ri-robot-line', tone: 'violet' },
-    { id: 'courses', label: '전체 강의', value: String(demoCourses.length), hint: '시스템에 등록된 강의 수', icon: 'ri-book-shelf-line', tone: 'amber' },
-    { id: 'users', label: '전체 사용자', value: String(demoUsers.length), hint: '운영 중인 학습자와 관리자', icon: 'ri-team-line', tone: 'indigo' },
-  ],
-  recent_activities: [
-    {
-      id: 'act_demo_1',
-      type: 'ai_summary',
-      title: 'AI 요약 생성 완료',
-      detail: 'AI 서비스 설계 입문 1주차 요약이 생성되었습니다.',
-      timestamp: now,
-      icon: 'ri-sparkling-line',
-      tone: 'violet',
-      course_id: demoCourses[0].id,
-      course_title: demoCourses[0].title,
-      lecture_id: demoCourseDetail.lectures[0].id,
-      lecture_title: demoCourseDetail.lectures[0].title,
-    },
-    {
-      id: 'act_demo_2',
-      type: 'shortform',
-      title: '숏폼 편집 완료',
-      detail: '전사 구간 기반 숏폼 초안이 내보내기 대기 상태입니다.',
-      timestamp: now,
-      icon: 'ri-scissors-cut-line',
-      tone: 'indigo',
-      course_id: demoCourses[2].id,
-      course_title: demoCourses[2].title,
-      lecture_id: 'lec_demo_sf_1',
-      lecture_title: '숏폼 컷 편집',
-    },
-    {
-      id: 'act_demo_3',
-      type: 'enrollment',
-      title: '신규 수강 등록',
-      detail: '데이터 기반 의사결정 강의에 학습자가 추가되었습니다.',
-      timestamp: now,
-      icon: 'ri-user-add-line',
-      tone: 'emerald',
-      course_id: demoCourses[1].id,
-      course_title: demoCourses[1].title,
-    },
-  ],
-  next_action: '전사와 숏폼 배포가 연결된 강의를 먼저 점검해 보세요.',
-};
-
-export const demoInsights: AIInsights = {
-  role: 'ADMIN',
-  summary: {
-    total_requests: 148,
-    success_rate: 0.96,
-    avg_latency_ms: 1240,
-    unique_users: 28,
-    recent_window_days: 7,
-  },
-  feature_stats: [
-    { feature: 'summary', label: '요약', count: 52, success_rate: 0.98, avg_latency_ms: 980 },
-    { feature: 'quiz', label: '퀴즈', count: 22, success_rate: 0.95, avg_latency_ms: 1410 },
-    { feature: 'smart', label: '스마트 챗', count: 41, success_rate: 0.96, avg_latency_ms: 1260 },
-  ],
-  intent_stats: [
-    { intent: 'lecture_explain', label: '개념 설명', count: 44, success_rate: 0.97, avg_confidence: 0.91 },
-    { intent: 'lecture_summary', label: '요약 요청', count: 33, success_rate: 0.98, avg_confidence: 0.93 },
-    { intent: 'quiz_generate', label: '퀴즈 생성', count: 18, success_rate: 0.94, avg_confidence: 0.88 },
-  ],
-  role_insight: {
-    role: 'ADMIN',
-    total_users: demoUsers.length,
-    published_courses: demoCourses.length,
-    total_enrollments: 3,
-    ai_usage_7d: 148,
-    feature_stats: [
-      { feature: 'summary', label: '요약', count: 52, success_rate: 0.98, avg_latency_ms: 980 },
-      { feature: 'smart', label: '스마트 챗', count: 41, success_rate: 0.96, avg_latency_ms: 1260 },
-    ],
-    intent_stats: [
-      { intent: 'lecture_explain', label: '개념 설명', count: 44, success_rate: 0.97, avg_confidence: 0.91 },
-      { intent: 'quiz_generate', label: '퀴즈 생성', count: 18, success_rate: 0.94, avg_confidence: 0.88 },
-    ],
-  },
-};
-
-export const demoLectureTranscript: LectureTranscript = {
-  id: 'trp_demo_ai_1',
-  lecture_id: demoLectureDetail.id,
-  user_id: demoUsers[3].id,
-  language: 'ko',
-  full_text: 'AI는 단순 자동완성이 아니라, 제품 흐름을 따라 기능을 연결하는 실행 파트너로 다뤄야 합니다. 전사와 타임스탬프는 복습을 더 빠르게 만들고, 숏폼은 학습 진입 장벽을 낮춥니다.',
-  segments: [
-    { index: 0, start_ms: 0, end_ms: 12000, text: 'AI는 단순 자동완성이 아니라, 제품 흐름을 따라 기능을 연결하는 실행 파트너로 다뤄야 합니다.' },
-    { index: 1, start_ms: 12000, end_ms: 26000, text: '전사와 타임스탬프는 복습을 더 빠르게 만들고,' },
-    { index: 2, start_ms: 26000, end_ms: 38000, text: '숏폼은 학습 진입 장벽을 낮춥니다.' },
-  ],
-  word_count: 28,
-  duration_ms: 38000,
-  stt_provider: 'cloudflare',
-  stt_model: 'whisper-large-v3',
-  created_at: now,
-};
-
-export const demoAudioExtraction: AudioExtraction = {
-  id: 'ext_demo_ai_1',
-  lecture_id: demoLectureDetail.id,
-  user_id: demoUsers[0].id,
-  source_type: 'video',
-  source_url: demoLectureDetail.video_url,
-  source_video_key: demoLectureDetail.video_asset_key,
-  source_video_name: 'ai-orchestration-intro.mp4',
-  source_content_type: 'video/mp4',
-  source_size_bytes: 28_642_124,
-  language: 'ko',
-  requested_stt_provider: 'cloudflare',
-  requested_stt_model: 'whisper-large-v3',
-  processing_job_id: 'job_demo_001',
-  processing_error: null,
-  processing_stage: 'completed',
-  processing_step: 'callback_done',
-  audio_url: 'https://cdn.mywayclass.dev/demo/ai-orchestration-intro-audio.mp3',
-  audio_format: 'mp3',
-  audio_duration_ms: 38000,
-  sample_rate: 48000,
-  channels: 2,
-  status: 'COMPLETED',
-  transcript_id: demoLectureTranscript.id,
-  stt_status: 'COMPLETED',
-  created_at: now,
-  processed_at: now,
-  updated_at: now,
-};
-
-export const demoLecturePipeline: LecturePipeline = {
-  lecture_id: demoLectureDetail.id,
-  transcript_status: 'COMPLETED',
-  summary_status: 'COMPLETED',
-  audio_status: 'COMPLETED',
-  transcript_id: demoLectureTranscript.id,
-  note_id: 'note_demo_001',
-  extraction_id: demoAudioExtraction.id,
-  updated_at: now,
-};
-
-export const demoMediaProcessorHealth: MediaProcessorHealth = {
-  ok: true,
-  public_base_url: 'https://media-demo.mywayclass.dev',
-  work_dir: '/workspace/media-processor',
-  token_configured: true,
-  callback_secret_configured: true,
-  ffmpeg: {
-    available: true,
-    path: '/usr/bin/ffmpeg',
-    version: '6.1',
-    output: 'ffmpeg version 6.1',
-  },
-  jobs: {
-    total: 12,
-    processing: 1,
-    completed: 10,
-    failed: 1,
-  },
-  recent_jobs: [
-    {
-      id: 'job_demo_001',
-      lecture_id: demoLectureDetail.id,
-      status: 'COMPLETED',
-      created_at: now,
-      updated_at: now,
-      audio_url: demoAudioExtraction.audio_url ?? null,
-      error_message: null,
-      stage: 'completed',
-      step: 'callback_done',
-      callback_status: 200,
-    },
-  ],
-  updated_at: now,
-};
-
+export { demoUsers, demoCourses } from './demo-auth-courses';
+export { demoCourseDetail, demoLectureDetail } from './demo-learning';
+export { demoDashboard, demoInsights } from './demo-analytics';
+export {
+  demoLectureTranscript,
+  demoAudioExtraction,
+  demoLecturePipeline,
+  demoMediaProcessorHealth,
+} from './demo-media';
